Update and rename login.js to auth.js
Update and rename login.js to auth.js
