--- controllers/login.js
@@ -1,309 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const fileUpload = require("express-fileupload");
-const jwt = require("jsonwebtoken");
-const bcrypt = require("bcrypt");
-
-const UserModel = require("../models/User");
-
-const key = {
-  tokenKey: "djghhhhuuwiwuewieuwieuriwu_cus"
-};
-
-const app = express();
-app.use(fileUpload());
-
-
-module.exports = {
-    register : async function(req, res) {
-        try {
-          const userExist = await UserModel.findOne({ email: req.body.email });
-      
-          if (!userExist) {
-            const hash = await bcrypt.hash(req.body.password, 8);
-            const User = new UserModel();
-      
-            User.username = req.body.userName;
-            User.password = hash;
-            // User.image = file.name;
-            User.email = req.body.email;
-      
-            const userCreate = await User.save();
-      
-            return res.json({
-              code: 200,
-              message: "Bạn đã đăng ký thành công",
-              data: { userCreate }
-            });
-          } else {
-            return res.json({
-              code: 200,
-              message: "Người dùng đã tồn tại",
-              data: null
-            });
-          }
-        } catch (err) {
-          return res.json({ code: 400, message: err.message, data: null });
-        }
-      },
-      login : async function(req, res, next) {
-        try {
-          const user = await UserModel.findOne({ email: req.body.email });
-      
-          if (!user) {
-            return res.json({
-              code: 401,
-              message: "Nhập sai email đăng nhập",
-              data: null
-            });
-          }
-      
-          if (user.isDelete === true) {
-            return res.json({
-              code: 401,
-              message: "Tài khoản của bạn đã bị khóa. Vui lòng liên hệ admin",
-              data: null
-            });
-          }
-      
-          const result = await bcrypt.compare(req.body.password, user.password);
-      
-          if (result) {
-            const token = jwt.sign({ _id: user._id }, key.tokenKey);
-      
-            const { username, role, email, image, _id } = user;
-      
-            return res.header("auth-token", token).json({
-              code: 200,
-              message: "Đăng nhập thành công",
-              data: { username, role, email, image, _id },
-              token
-            });
-          } else {
-            return res.json({
-              code: 400,
-              message: "Nhập sai password",
-              data: null
-            });
-          }
-        } catch (err) {
-          return res.json({ code: 400, message: err.message, data: null });
-        }
-      },
-      get_user : async (req, res) => {
-        const userToken = req.params.token;
-      
-        try {
-          let decoded = jwt.decode(userToken);
-      
-          if (decoded) {
-            const id = decoded._id;
-            const user = await UserModel.findOne({ _id: id });
-            const { username, role, email, image, _id } = user;
-      
-            return res.json({
-              code: 200,
-              data: { username, role, email, image, _id }
-            })
-      
-          } else {
-            return res.json({
-              message: "Lỗi xác thực!"
-            })
-          }
-      
-        } catch (err) {
-          return res.json({
-            code: 400,
-            message: err
-          })
-        }
-    },
-    updateName : async (req, res) => {
-        const userExist = await UserModel.findOne({ _id: req.params.id });
-      
-        if (userExist) {
-          try {
-            const user = {
-              username: req.body.userName
-            };
-      
-            const updateUserName = await UserModel.findOneAndUpdate(
-              { _id: req.params.id },
-              user,
-              {
-                new: true,
-                useFindAndModify: false
-              }
-            );
-      
-            if (updateUserName) {
-              res.json({
-                data: updateUserName,
-                code: 200,
-                message: "Cập nhật name thành công"
-              });
-            }
-          } catch (error) {
-            res.status(500).json({
-              message: error
-            });
-          }
-        }
-      },
-      updateEmail : async (req, res) => {
-        const userExist = await UserModel.findOne({ _id: req.params.id });
-        const emailExist = await UserModel.findOne({ email: req.body.email });
-      
-        if (userExist) {
-          if (emailExist) {
-            res.json({
-              code: 404,
-              message: "Email đã tồn tại"
-            });
-          } else {
-            try {
-              const user = {
-                email: req.body.email
-              };
-      
-              const updateUserEmail = await UserModel.findOneAndUpdate(
-                { _id: req.params.id },
-                user,
-                {
-                  new: true,
-                  useFindAndModify: false
-                }
-              );
-      
-              if (updateUserEmail) {
-                res.json({
-                  code: 200,
-                  data: updateUserEmail,
-                  message: "Cập nhật email thành công"
-                });
-              }
-            } catch (error) {
-              res.status(500).json({
-                message: error
-              });
-            }
-          }
-        }
-      },
-      updatePassword : async (req, res) => {
-        const userExist = await UserModel.findOne({ _id: req.params.id });
-      
-        if (userExist) {
-          const comparePassword = await bcrypt.compare(
-            req.body.currentPassword,
-            userExist.password
-          );
-      
-          if (comparePassword) {
-            try {
-              const hashPassword = await bcrypt.hash(req.body.newPassword, 8);
-      
-              const user = {
-                password: hashPassword
-              };
-      
-              const updateUserPassword = await UserModel.findOneAndUpdate(
-                { _id: req.params.id },
-                user,
-                {
-                  new: true,
-                  useFindAndModify: false
-                }
-              );
-      
-              if (updateUserPassword) {
-                res.json({
-                  data: updateUserPassword,
-                  code: 200,
-                  message: "Cập nhật password thành công"
-                });
-              }
-            } catch (error) {
-              res.status(500).json({
-                message: error
-              });
-            }
-          } else {
-            res.json({
-              code: 404,
-              message: "Nhập password hiện tại không đúng"
-            });
-          }
-        }
-      },
-      uploadAvatar : async (req, res) => {
-        const userExist = await UserModel.findOne({ _id: req.params.id });
-      
-        if (userExist) {
-          const file = req.files.file;
-      
-          if (file) {
-            try {
-              file.mv(`${__dirname}/../client/public/uploads/users/${file.name}`);
-      
-              const user = {
-                image: file.name
-              };
-      
-              const updateUserAvatar = await UserModel.findOneAndUpdate(
-                { _id: req.params.id },
-                user,
-                {
-                  new: true,
-                  useFindAndModify: false
-                }
-              );
-      
-              if (updateUserAvatar) {
-                res.json({
-                  data: updateUserAvatar,
-                  code: 200,
-                  message: "Thay đổi avatar thành công"
-                });
-              }
-            } catch (error) {
-              res.json({
-                code: 500,
-                message: "Thay đổi avatar thất bại"
-              });
-            }
-          }
-        }
-      },
-      checkToken : (req, res) => {
-        const token = req.body.token;
-        const role = req.body.role;
-        const tokenAuth = key.tokenKey;
-      
-        jwt.verify(tokenAuth, token, function(err, decoded) {
-          if (err) {
-            res.json({
-              role: role,
-              message: "Error"
-            });
-          }
-      
-          if (decoded && role === "admin") {
-            res.json({
-              role: role
-            });
-          }
-          if (decoded && role === "editor") {
-            res.json({
-              role: role
-            });
-          }
-          if (decoded && role === "customer") {
-            res.json({
-              role: role
-            });
-          }
-        });
-      }
-}
